Refactor chat options screens: Migrate GalleryOption, GroupOption, ConversationName, SearchFooter, SearchHeader, and DescriptionOption to new structure; remove deprecated files and implement new components for better organization and maintainability.
