+++ app/App.js
@@ -21,7 +21,7 @@ import ReceivedFriendRequests from "./screens/directory/friend/friendRequest/Rec
 import UserProfile from "./screens/profile/userProfile/UserProfile"; // Đảm bảo đường dẫn đúng
 import AddFriend from "./screens/profile/userProfile/AddFriend"; // Đảm bảo đường dẫn đúng
 import AcceptFriend from "./screens/profile/userProfile/AcceptFriend"; // Đảm bảo đường dẫn đúng
-import Options from "./screens/optional/OptionalScreen";
+import Options from "./screens/chat/optional/OptionalScreen";
 import { NavigationContainer } from "@react-navigation/native";
 import Verify from "./screens/register/Verify";
 import VerifyOTPCode from "./screens/register/verifyOTPCode/VerifyOTPCode";
@@ -44,79 +44,78 @@ import EnterAvatar from "./screens/register/enterAvatar/EnterAvatar";
 import LoginByQR from "./screens/login/LoginByQR";
 import { SocketProvider } from "./screens/socket/SocketContext";
 
-
-
 const Stack = createNativeStackNavigator();
 
 function RootLayout() {
   return (
     <AuthProvider>
       <SocketProvider>
-      <NavigationContainer>
-        <Stack.Navigator
-          screenOptions={{
-            headerShown: false,
-            animation: "fade", // Tắt hoạt ảnh chuyển cảnh
-          }}
-        >
-          <Stack.Screen name="AuthLoading" component={AuthLoading} />
+        <NavigationContainer>
+          <Stack.Navigator
+            screenOptions={{
+              headerShown: false,
+              animation: "fade", // Tắt hoạt ảnh chuyển cảnh
+            }}
+          >
+            <Stack.Screen name="AuthLoading" component={AuthLoading} />
 
-          <Stack.Screen name="Main" component={Main} />
-          <Stack.Screen name="Register" component={Register} />
-          <Stack.Screen name="VerifyOTPCode" component={VerifyOTPCode} />
-          <Stack.Screen name="EnterName" component={EnterName} />
-          <Stack.Screen name="EnterInfo" component={EnterInfo} />
-          <Stack.Screen name="Login" component={Login} />
-          <Stack.Screen name="ForgotPassword" component={ForgotPassword} />
-          <Stack.Screen
-            name="SocialNetworkTerms"
-            component={SocialNetworkTerms}
-          />
-          <Stack.Screen name="CameraScreen" component={CameraScreen} />
-          {/* <Stack.Screen name="TermsOfService" component={TermsOfService} /> */}
-          <Stack.Screen name="Home" component={Home} />
-          <Stack.Screen name="Chat" component={Chat} />
-          <Stack.Screen name="Profile" component={Profile} />
-          <Stack.Screen name="Discover" component={Discover} />
-          <Stack.Screen name="Diary" component={Diary} />
-          <Stack.Screen name="MyProfile" component={MyProfile} />
-          <Stack.Screen name="Setting" component={Setting} />
-          <Stack.Screen
-            name="ReceivedFriendRequests"
-            component={ReceivedFriendRequests}
-          />
-          <Stack.Screen name="UserProfile" component={UserProfile} />
-          <Stack.Screen name="AddFriend" component={AddFriend} />
-          <Stack.Screen name="AcceptFriend" component={AcceptFriend} />
-          <Stack.Screen
-            name="Options"
-            component={Options}
-            options={{ animation: "slide_from_right" }}
-          />
-          <Stack.Screen name="Verify" component={Verify} />
-          <Stack.Screen name="ScanQR" component={ScanQR} />
-          <Stack.Screen name="VerificationCode" component={VerificationCode} />
-          <Stack.Screen
-            name="CreateNewPassword"
-            component={CreateNewPassword}
-          />
-          <Stack.Screen
-            name="AccountAndSecurity"
-            component={AccountAndSecurityScreen}
-          />
-          <Stack.Screen name="Personal" component={Personal} />
-          <Stack.Screen name="Edit" component={Edit} />
-          <Stack.Screen name="UpdatePassword" component={UpdatePassword} />
-          <Stack.Screen
-            name="EnterAvatar"
-            component={EnterAvatar}
-            options={{ animation: "slide_from_right" }}
-          />
-          <Stack.Screen name="LoginByQR" component={LoginByQR} />
-
-
-        </Stack.Navigator>
-      </NavigationContainer>
+            <Stack.Screen name="Main" component={Main} />
+            <Stack.Screen name="Register" component={Register} />
+            <Stack.Screen name="VerifyOTPCode" component={VerifyOTPCode} />
+            <Stack.Screen name="EnterName" component={EnterName} />
+            <Stack.Screen name="EnterInfo" component={EnterInfo} />
+            <Stack.Screen name="Login" component={Login} />
+            <Stack.Screen name="ForgotPassword" component={ForgotPassword} />
+            <Stack.Screen
+              name="SocialNetworkTerms"
+              component={SocialNetworkTerms}
+            />
+            <Stack.Screen name="CameraScreen" component={CameraScreen} />
+            {/* <Stack.Screen name="TermsOfService" component={TermsOfService} /> */}
+            <Stack.Screen name="Home" component={Home} />
+            <Stack.Screen name="Chat" component={Chat} />
+            <Stack.Screen name="Profile" component={Profile} />
+            <Stack.Screen name="Discover" component={Discover} />
+            <Stack.Screen name="Diary" component={Diary} />
+            <Stack.Screen name="MyProfile" component={MyProfile} />
+            <Stack.Screen name="Setting" component={Setting} />
+            <Stack.Screen
+              name="ReceivedFriendRequests"
+              component={ReceivedFriendRequests}
+            />
+            <Stack.Screen name="UserProfile" component={UserProfile} />
+            <Stack.Screen name="AddFriend" component={AddFriend} />
+            <Stack.Screen name="AcceptFriend" component={AcceptFriend} />
+            <Stack.Screen
+              name="Options"
+              component={Options}
+              options={{ animation: "slide_from_right" }}
+            />
+            <Stack.Screen name="Verify" component={Verify} />
+            <Stack.Screen name="ScanQR" component={ScanQR} />
+            <Stack.Screen
+              name="VerificationCode"
+              component={VerificationCode}
+            />
+            <Stack.Screen
+              name="CreateNewPassword"
+              component={CreateNewPassword}
+            />
+            <Stack.Screen
+              name="AccountAndSecurity"
+              component={AccountAndSecurityScreen}
+            />
+            <Stack.Screen name="Personal" component={Personal} />
+            <Stack.Screen name="Edit" component={Edit} />
+            <Stack.Screen name="UpdatePassword" component={UpdatePassword} />
+            <Stack.Screen
+              name="EnterAvatar"
+              component={EnterAvatar}
+              options={{ animation: "slide_from_right" }}
+            />
+            <Stack.Screen name="LoginByQR" component={LoginByQR} />
+          </Stack.Navigator>
+        </NavigationContainer>
       </SocketProvider>
     </AuthProvider>
   );
